+++ src/pages/home/subPages/components/memo/hook/index.ts
@@ -1,1 +1,3 @@
-export { default as useMemoManager } from './useMemoManager';
+export { default as useCreateMemoManager } from './useCreateMemoManager';
+export { default as useUpdateMemoManager } from './useUpdateMemoManger';
+export { default as useDeleteMemoManager } from './useDeleteMemoManager';
