--- src/pages/home/subPages/create/hook/index.ts
@@ -1,1 +1,0 @@
-export { default as useCreateMemoManager } from './useCreateMemoManager';
